Modal One rewrite
Now that the UI has changed I wanted to update the images. I am also using png files for the images rather than the gif files. The focus of this this first modal is to demonstrate what every day practice looks like for schools that use reVIEW talent Feedback System.